Remove Flyway and use Primsa Migrate
DESCRIPTION
Prisma is used as migration tool because their tool uses Prisma schema as the source of truth.
Flyway is removed so Prisma will gurantee to handle all interactions with the DB.
Do note that Prisma migrate is a preview feature are not to be use for PROD.
The better approach is to use Flyway and maybe switch to Prisma Migrate once it becomes official.

CHANGES
[x] Remove Flyway integration
[x] Add Prisma Migrate
[x] Add the base migration scripting
[x] Add seeding script
[x] Add npm scripts for these and hook them up